refactor: export as named export to Header components
