+++ apps/blog/src/components/Header/index.tsx
@@ -1,0 +1,2 @@
+export { MainHeader } from './MainHeader';
+export { PostHeader } from './PostHeader';
