+++ apps/blog/src/pages/[slug].tsx
@@ -5,7 +5,7 @@ import { PageProgressBar } from 'core';
 import AuthorSection from '../components/AuthorSection';
 import Comments from '../components/Comments';
 import DateAndCategoryLink from '../components/DateAndCategoryLink';
-import PostHeader from '../components/Header/PostHeader';
+import { PostHeader } from '../components/Header';
 import SEO from '../components/SEO';
 import TOC from '../components/TOC';
 import { getAllPosts } from '../libs/api';
