+++ apps/blog/src/pages/category/[category].tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { useTheme } from '@nextui-org/react';
 
 import AuthorSection from '../../components/AuthorSection';
-import MainHeader from '../../components/Header/MainHeader';
+import { MainHeader } from '../../components/Header';
 import PostCard from '../../components/PostCard';
 import SEO from '../../components/SEO';
 import useInfiniteScroll from '../../hooks/useInfiniteScroll';
