+++ apps/blog/src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { useTheme } from '@nextui-org/react';
 
 import AuthorSection from '../components/AuthorSection';
-import MainHeader from '../components/Header/MainHeader';
+import { MainHeader } from '../components/Header';
 import PostCard from '../components/PostCard';
 import SEO from '../components/SEO';
 import useInfiniteScroll from '../hooks/useInfiniteScroll';
